third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-setLocalDescription-parameterless.https_interop-2026-expected.txt

This is a testharness.js-based test.
[FAIL] RTCSessionDescription constructed without type throws TypeError
  assert_equals: expected "TypeError" but got "ReferenceError"
Harness: the test ran to completion.
